# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/es/get/": {
            "get": {
                "description": "根据id获取对象，可以是author，work，institution,venue,concept W4237558494,W2009180309,W2984203759",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据id获取对象 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "userid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"res\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"id type error\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"status\":201,\"msg\":\"es get err or not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/es/getAuthorRelationNet": {
            "get": {
                "description": "根据author的id获取专家关系网络, 目前会返回Top N的关系网，N=10，后续可以讨论修改N的大小或者传参给我\n\n目前接口时延约为1s, 后续考虑把计算出来的结果存入数据库，二次查询时延降低\n",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据author的id获取专家关系网络 hr",
                "parameters": [
                    {
                        "enum": [
                            "A2764814280",
                            "A2900471938",
                            "A2227665069"
                        ],
                        "type": "string",
                        "description": "author_id",
                        "name": "author_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{ \"Vertex_set\":[], \"Edge_set\":[]}}",
                        "schema": {
                            "$ref": "#/definitions/response.AuthorRelationNet"
                        }
                    },
                    "201": {
                        "description": "{\"msg\":\"Get Author Relation Net Error\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/prefix": {
            "post": {
                "description": "根据前缀得到搜索建议，返回results 字符串数组",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据前缀得到搜索建议，返回results 字符串数组 hr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field 表示需要查询的字段名",
                        "name": "Field",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Prefix 表示用户已经输入的前缀",
                        "name": "Prefix",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"msg\": \"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"msg\": 参数错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"success\": false, \"msg\": \"es服务出错\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/advanced": {
            "post": {
                "description": "高级搜索，query是一个map列表， 每个map包含\"content\" \"field\" \"logic\"\nlogic 仅包含[\"and\", \"or\", \"not\"]\nfield 仅包含[\"title\", \"abstract\", \"venue\", \"publisher\", \"author\", \"institution\", \"concept\"]\n对于年份的筛选，在query里面 field是\"publication_date\" logic默认为and， 该map下有\"begin\" \"end\"分别是开始和结束\nsort=0为默认排序（降序） =1为按引用数降序 =2按发表日期由近到远\nasc=0为降序 =1为升序\n{ \"asc\": false,\"conds\": {\"venue\":\"International Journal for Research in Applied Science and Engineering Technology\",\"author\": \"Zenith Nandy\"},\"page\": 1,\"query\": [{\"field\": \"title\",\"content\": \"python\",\"logic\": \"and\"},{\"field\": \"publication_date\",\"begin\": \"2021-12-01\",\"end\":\"2022-06-01\",\"logic\": \"and\"}],\"size\": 8,\"sort\": 0}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "esSearch"
                ],
                "summary": "高级搜索 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.AdvancedSearchQ"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/es/search/author": {
            "get": {
                "description": "根据作者名字搜索作者,模糊搜索",
                "tags": [
                    "esSearch"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query_word",
                        "name": "query_word",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sort",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "asc",
                        "name": "asc",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{},\"msg\": \"Author Search Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"Author Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/author2": {
            "get": {
                "description": "根据作者名字搜索作者 via openalex",
                "tags": [
                    "esSearch"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query_word",
                        "name": "query_word",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort=cited_by_count|...:_|desc|asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{},\"msg\": \"Author Search Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"openalex Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"openalex Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/base": {
            "post": {
                "description": "基本搜索，Cond里面填筛选条件，key仅包含[\"type\", \"author\", \"institution\", \"publisher\", \"venue\", \"publication_year\"]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "esSearch"
                ],
                "summary": "基本搜索 txc",
                "parameters": [
                    {
                        "description": "搜索条件",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.BaseSearchQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"res\":{obeject}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "{\"status\":201,\"err\":\"es search err\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/doi": {
            "post": {
                "description": "使用doi查找work，未测试，请勿使用",
                "tags": [
                    "esSearch"
                ],
                "summary": "使用doi查找work，未测试，请勿使用 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "doi",
                        "name": "doi",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "response.AdvancedSearchQ": {
            "type": "object",
            "properties": {
                "asc": {
                    "type": "boolean"
                },
                "conds": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "query": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "response.AuthorRelationNet": {
            "type": "object",
            "properties": {
                "edge_set": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "vertex_set": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            }
        },
...
    }
}